fixup! feat: improve template code
