+++ .github/workflows/ci.yml
@@ -35,9 +35,6 @@ jobs:
       - name: Install Python dependencies
         run: pip install -r requirements/dev.txt
 
-      - name: Install Python dependencies
-        run: pip install -r requirements/dev.txt
-
       - name: Python test and coverage
         run: pytest
